fix: Add alternate generator based on thumbnail
