+++ packages/static-wado-util/lib/index.js
@@ -33,6 +33,7 @@ module.exports.program = program;
 module.exports.Stats = Stats;
 module.exports.getStudyUIDPathAndSubPath = require("./getStudyUIDPathAndSubPath");
 module.exports.createStudyDirectories = require("./createStudyDirectories");
+module.exports.readBulkDataValue = require("./reader/readBulkDataValue");
 
 module.exports.default = {
   ...module.exports,
